+++ src/components/inputs/multiline-text-input/multiline-text-input.js
@@ -7,7 +7,7 @@ import { css } from '@emotion/core';
 import FlatButton from '../../buttons/flat-button';
 import { AngleUpIcon, AngleDownIcon } from '../../icons';
 import filterDataAttributes from '../../../utils/filter-data-attributes';
-import useCollapsible from '../../use-collapsible';
+import useCollapsible from '../../../hooks/use-collapsible';
 import Spacings from '../../spacings';
 import Constraints from '../../constraints';
 import messages from './messages';
